Added SDLSoundManager class + working sound
- implemented SDLSoundManager
- SDL_main is now in charge of the list of sounds to load
- SDL_main is now in charge of the Sound_Play function implementation
- added jump, land, pickup, splat, and screen transition sounds
- added sounds to visual studio project and make them copy on build

+++ game/game_types.h
@@ -49,7 +49,7 @@ typedef struct GameData
 	u8 string_highScore[SCORE_STRING_SIZE];
 
 	u32 highScore;
-	u32 paused;
+	u8 paused;
 } GameData;
 
 #endif
